Server/requirements.txt
asgiref==3.8.1
blosc==1.11.1
channels==4.1.0
cryptography==42.0.7
Django==4.2.12
djangorestframework==3.15.1
djangorestframework_simplejwt==5.3.1
neomodel==5.3.0
numpy==1.26.4
pandas==2.2.2
python_dateutil==2.9.0.post0
pytz==2024.1
Requests==2.31.0
scipy==1.13.0
